Update backend files and vercel config

+++ Automation/RemoveUnverifiedUser.js
@@ -3,6 +3,11 @@ const User = require("../Model/UserModel");
 
 const removeUnverifiedUsers = () => {
   const deleteUnverified = async () => {
+    if (mongoose.connection.readyState !== 1) {
+      console.log("⚠️ MongoDB not connected yet. Skipping delete.");
+      return;
+    }
+
     try {
       const thirtyMinutesAgo = new Date(Date.now() - 30 * 60 * 1000);
       const result = await User.deleteMany({
@@ -14,7 +19,8 @@ const removeUnverifiedUsers = () => {
       console.error("❌ Error deleting unverified users:", error);
     }
   };
-  deleteUnverified();
+
+  // Schedule cron job to run every 30 minutes
   cron.schedule("*/30 * * * *", deleteUnverified);
   console.log(
     "⏰ Cron job to remove unverified users scheduled every 30 minutes."
